phase-3-ai-chatbot/backend/requirements.txt
# Phase III AI-Powered Chatbot Backend Dependencies
# Date: 2026-01-15

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlmodel>=0.0.14
asyncpg>=0.29.0

# AI Providers
openai>=1.10.0
google-genai>=0.1.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# MCP SDK (Model Context Protocol)
mcp>=1.25.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
slowapi>=0.1.9
bleach>=6.1.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.1
ruff>=0.1.14
